se crea la ruta updatem falta ajustar que use el id que viene en el parametro

+++ src/main/java/com/aplication/rest/SpringBootRest/controllers/ProductController.java
@@ -3,7 +3,6 @@ package com.aplication.rest.SpringBootRest.controllers;
 import com.aplication.rest.SpringBootRest.controllers.dto.ProductDTO;
 import com.aplication.rest.SpringBootRest.entities.Product;
 import com.aplication.rest.SpringBootRest.service.IProductService;
-import org.hibernate.sql.results.internal.ResolvedSqlSelection;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -66,29 +65,14 @@ public class ProductController {
         return ResponseEntity.created(new URI("api/product/save")).build();
     }
 
-    @PutMapping ("/update/{id}")
-    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO){
-         Optional<Product> productOptional = productService.findById(id);
-
-         if(productOptional.isPresent() ){
-             Product product = productOptional.get();
-             product.setName(productDTO.getName());
-             productService.save(product);
-             return ResponseEntity.ok("Registro Actualizado");
-         }
-         return ResponseEntity.notFound().build();
-    }
-
     @GetMapping("/get/{id}")
     public ResponseEntity<ProductDTO> getByid(@PathVariable Long id) {
         if (id != null && productService.findById(id).isPresent() ){
             productService.getById(id);
             return ResponseEntity.ok(productService.getById(id));
         }
-        return ResponseEntity.badRequest().build();
+            return ResponseEntity.badRequest().build();
     }
-
-
 
     @GetMapping("/getAll")
     public ResponseEntity<?> getAll() {
@@ -101,11 +85,41 @@ public class ProductController {
         ProductDTO product = productService.create(productDTO);
         return ResponseEntity.ok(product);
     }
+
     @DeleteMapping ("/delete/{id}")
     public ResponseEntity<?> deleteById (@PathVariable Long id) {
-        productService.deleteById(id);
-        return ResponseEntity.ok("Bien hecho");
+       if(productService.findById(id).isPresent()){
+           productService.deleteById(id);
+           productService.otraFuncion();
+           return ResponseEntity.ok("Registro Eliminado");
+       }
+           return ResponseEntity.badRequest().build();
+       }
+
+       @PutMapping("/updatem/{id}")
+       public ResponseEntity<?> updateM (@PathVariable Long id, @RequestBody ProductDTO productDTO){
+              Optional<Product> productOptional = productService.findById(id);
+
+              if(productOptional.isPresent() ){
+                  productService.updatem(productDTO);
+                  return ResponseEntity.ok("Registro Actualizado");
+              }
+              return ResponseEntity.badRequest().build();
+       }
+
+    @PutMapping ("/update/{id}")
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO){
+        Optional<Product> productOptional = productService.findById(id);
+
+        if(productOptional.isPresent() ){
+            Product product = productOptional.get();
+            product.setName(productDTO.getName());
+            productService.save(product);
+            return ResponseEntity.ok("Registro Actualizado");
+        }
+        return ResponseEntity.notFound().build();
     }
+
 }
